Move new classes to streamer package
