+++ hudi-utilities/src/main/java/org/apache/hudi/utilities/streamer/ConfigurationHotUpdateStrategy.java
@@ -7,16 +7,17 @@
  * "License"); you may not use this file except in compliance
  * with the License.  You may obtain a copy of the License at
  *
- *      http://www.apache.org/licenses/LICENSE-2.0
+ *   http://www.apache.org/licenses/LICENSE-2.0
  *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
  */
 
-package org.apache.hudi.utilities.deltastreamer;
+package org.apache.hudi.utilities.streamer;
 
 import org.apache.hudi.common.config.TypedProperties;
 import org.apache.hudi.common.util.Option;
@@ -26,10 +27,10 @@ import org.apache.hudi.common.util.Option;
  */
 public abstract class ConfigurationHotUpdateStrategy {
 
-  protected HoodieDeltaStreamer.Config cfg;
+  protected HoodieStreamer.Config cfg;
   protected TypedProperties properties;
 
-  public ConfigurationHotUpdateStrategy(HoodieDeltaStreamer.Config cfg,
+  public ConfigurationHotUpdateStrategy(HoodieStreamer.Config cfg,
                                         TypedProperties properties) {
     this.cfg = cfg;
     this.properties = properties;
