+++ hudi-utilities/src/main/java/org/apache/hudi/utilities/streamer/ConfigurationHotUpdateStrategyUtils.java
@@ -7,16 +7,17 @@
  * "License"); you may not use this file except in compliance
  * with the License.  You may obtain a copy of the License at
  *
- *      http://www.apache.org/licenses/LICENSE-2.0
+ *   http://www.apache.org/licenses/LICENSE-2.0
  *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
  */
 
-package org.apache.hudi.utilities.deltastreamer;
+package org.apache.hudi.utilities.streamer;
 
 import org.apache.hudi.common.config.TypedProperties;
 import org.apache.hudi.common.util.Option;
@@ -30,11 +31,11 @@ public class ConfigurationHotUpdateStrategyUtils {
    * Creates a {@link ConfigurationHotUpdateStrategy} class via reflection.
    *
    * <p>If the class name of {@link ConfigurationHotUpdateStrategy} is configured
-   * through the {@link HoodieDeltaStreamer.Config#configHotUpdateStrategyClass}.
+   * through the {@link HoodieStreamer.Config#configHotUpdateStrategyClass}.
    */
   public static Option<ConfigurationHotUpdateStrategy> createConfigurationHotUpdateStrategy(
       String strategyClass,
-      HoodieDeltaStreamer.Config cfg,
+      HoodieStreamer.Config cfg,
       TypedProperties properties) throws HoodieException {
     try {
       return StringUtils.isNullOrEmpty(strategyClass)
